Remove a workaround for an old problem
This approach of inserting arguments with secondary files at the start
of the inputs dates back to commit
77819d7005f428241edb39acff740546aa0156c9 (see line 57 of
CWL/GATKtools/helper_functions.py); apparently it was intended to work
around a "bug in CWL" (see line 186 of the same file at
e65df75d1e6f64e46dbed562171a29bebb86e478). The comment next to it was
updated to mention refIndex and refDict in commit
24d515c885bfddcf53f54d6e04abce28d638f9d1, but those two inputs are not
included in the generated CWL since gatk-cwl-generator 2.0, so I think
it's safe to remove this workaround at this point.